==> notebooks/1_extraction_moviesdb.ipynb <==
### 0. Imports

# Data transformation
# ---------------------------------------
import pandas as pd
import polars as pl

# API calls
# ---------------------------------------
import requests

# Scraping
# ---------------------------------------
from bs4 import BeautifulSoup


# Aynchronicity
# ---------------------------------------
import asyncio
import aiohttp


import json


import time

# 1. Introducción - Extracción datos Movies Database

Este notebook detalla la exploración acarreada para la extracción de datos a través de la API Movies Database.

# 2. Extraccion

- Películas desde 1990 hasta la actualidad.
- Géneros: Drama, Comedy, Action, Fantasy, Horror, Mystery, Romance, Thriller.
- Información necesaria:
  - Tipo (corto o película).
  - Nombre.
  - Año y mes de estreno.
  - ID de la película.

Request de ejemplo:

url = "https://moviesdatabase.p.rapidapi.com/titles"

querystring = {"genre":"Drama","startYear":"1990","titleType":"movie","page":"70","limit":"50"}

headers = {
	"x-rapidapi-key": "c09b699b97mshc664835db7a2ca3p1f803fjsn67c91d35ded1",
	"x-rapidapi-host": "moviesdatabase.p.rapidapi.com"
}

response = requests.get(url, headers=headers, params=querystring)

print(response.json())

Hay paginación. No se conoce el límite de páginas por género y tipo de cinta.

Dado que hay 1000 consultas por hora, que hay que hacer consultas separadas para 'short' y para 'movie', así como para los géneros, se puede tratar de intentar un approach rápido por asincronía de hacer todas las consultas en lugar usar un bucle que trate de encontrar el límite de páginas para cada uno, con el consiguiente tiempo de diseño de la lógica para capturarlo.

Extraer campos de requeridos de cada pelicula.

# page
response.json()
# id 
response.json()["results"][0]["id"]
# type
response.json()["results"][0]["titleType"]["id"]

# nombre
response.json()["results"][0]["titleText"]["text"]

# año y mes de estreno
response.json()["results"][0]["releaseDate"]["year"]
response.json()["results"][0]["releaseDate"]["month"]



# miguellv96@gmail.com
headers = {
    "x-rapidapi-key": "c09b699b97mshc664835db7a2ca3p1f803fjsn67c91d35ded1",
    "x-rapidapi-host": "moviesdatabase.p.rapidapi.com"
}


# miguel.lopezvirues@alum.uca.es
headers = {
    'x-rapidapi-key': "5027380e12msh3259c6d3e08bbe5p1b31bbjsn6f9e42d6d53b",
    'x-rapidapi-host': "moviesdatabase.p.rapidapi.com"
}

# miguel.lopezvirues@gmail.com
headers = {
    'x-rapidapi-key': "98ab80c541msh548e16e3c595c79p101433jsn6131790668a3",
    'x-rapidapi-host': "moviesdatabase.p.rapidapi.com"
}

# hackio projects 1
headers = {
    "x-rapidapi-key": "997763c3bamsheeaaefa4d7efc72p100dcfjsnc6ba0f0812da",
    "x-rapidapi-host": "moviesdatabase.p.rapidapi.com"
}

# hackio projects 2 
headers = {
	"x-rapidapi-key": "d42d6bd155mshfbdc1f0c83f26f5p12fca4jsnf8aa6552f0e5",
	"x-rapidapi-host": "moviesdatabase.p.rapidapi.com"
}

async def request_movies_async(querystring):
    
    url = "https://moviesdatabase.p.rapidapi.com/titles"

    headers = {
        "x-rapidapi-key": "c09b699b97mshc664835db7a2ca3p1f803fjsn67c91d35ded1",
        "x-rapidapi-host": "moviesdatabase.p.rapidapi.com"
    }

    async with aiohttp.ClientSession() as session:
        async with session.get(url, headers=headers, params=querystring) as response:
            if response.status == 500:
                print(f"Max operation time for querystring {querystring}. Retrying...")
                await asyncio.sleep(3)  # Wait for 3 seconds before retrying
                return await request_movies_async(querystring)  # Retry the same request
            elif response.status == 200:
                try:
                    results = await response.json()
                    pagina = results["page"]
                    results = results["results"]
                    for result in results:
                        result["pagina"] = pagina
                        result["genre"] = querystring["genre"]

                except KeyError as e:
                    print(f"KeyError with querystring {querystring}: {e}")
                    return []  # returning an empty list for consistency
                except TypeError as e:
                    print(f"TypeError with querystring {querystring}: {e}")
                    return []
                except Exception as e:
                    print(f"Unexpected error with querystring {querystring}: {e}")
                    return []
            else:
                print(f"Request failed with status {response.status} for {querystring}")
                return []
                raise ValueError(f"HTTP Error: {response.status}")
            

    return results

async def get_movies_info(querystring_list): 

    tasks = [request_movies_async(querystring) for querystring in querystring_list]
    results = await asyncio.gather(*tasks)

    return results

def create_querystrings(end_page, title_type_list, genres_list, start_year):
    querystring_list = []

    for page in range(1,end_page+1):
        for title_type in title_type_list:
            for genre in genres_list:
                base_querystring = {"genre":None,"startYear":str(start_year),"titleType":None,"page":None,"limit":"50","sort":"year.incr"}

                base_querystring["genre"] = genre
                base_querystring["titleType"] = title_type
                base_querystring["page"] = page

                querystring_list.append(base_querystring)

    return querystring_list
                



Creación de función de extracción de datos:

def generar_lista_tuplas(lista_jsons):

    lista_tuplas_peliculas = []
    for response in lista_jsons:
        # page
        page = response["pagina"]
        # id 
        id = response["id"]
        # type
        title_type = response["titleType"]["id"]

        # nombre

        title = response["titleText"]["text"]

        # año y mes de estreno

        try:
            
            year = response["releaseDate"]["year"]
            month = response["releaseDate"]["month"]
        except:

            year = response["releaseYear"]["year"]
            month = None

        # genre
        genre = response["genre"]

        pelicula_tupla = (page, id, title_type, title, genre, year, month)

        lista_tuplas_peliculas.append(pelicula_tupla)

    return lista_tuplas_peliculas



test_querystring_list = create_querystrings(1, ["movie"],["Drama"],1990)

results = await get_movies_info(test_querystring_list)

resultados_flat = [sublista for lista in results for sublista in lista]

generar_lista_tuplas(resultados_flat)

Realizar la llamada global:

genre_list = ["Drama", "Comedy", "Action", "Fantasy", "Horror", "Mystery", "Romance", "Thriller"]
title_type_list =  ["movie","short"]
end_page = 55
end_year = 1990

querystring_list = create_querystrings(end_page, title_type_list,genre_list,end_year)
print("Len of querystring list" ,len(querystring_list))
querystring_list[:10]

results_400 = await get_movies_info(querystring_list[:400])

results_480 = await get_movies_info(querystring_list[400:])

resultados_400_flat = [sublista for lista in results_400 for sublista in lista]

lista_tuplas_400 = generar_lista_tuplas(resultados_400_flat)
df_400 = pd.DataFrame(lista_tuplas_400)
df_400.to_csv("../data/movies_400.csv")
df_400

resultados_480_flat = [sublista for lista in results_480 for sublista in lista]

lista_tuplas_480 = generar_lista_tuplas(resultados_480_flat)
df_480 = pd.DataFrame(lista_tuplas_480)
df_480.to_csv("../data/movies_480.csv")

df_480

df_total = pd.concat([df_400, df_480],axis=0)
df_total.columns = ["pagina","id","tipo_titulo","titulo","genero","año","mes"]
df_total

df_total = df_total.drop_duplicates(subset="id")
df_total.to_csv("../data/movies_total.csv")
