==> notebooks/3_limpieza_datos.ipynb <==
### 0. Imports

# Data transformation
# ---------------------------------------
import pandas as pd
import polars as pl
import numpy as np

# API calls
# ---------------------------------------
import requests

# Scraping
# ---------------------------------------
from bs4 import BeautifulSoup


# Aynchronicity
# ---------------------------------------
import asyncio
import aiohttp


import json


import time


import re

from tqdm import tqdm

from joblib import Parallel, delayed

import sys

if sys.platform == 'win32':
    loop = asyncio.ProactorEventLoop()
    asyncio.set_event_loop(loop)

sys.path.append("..")

import src.soporte_eda as se

# 1. Introducción - Limpieza datos Movies Database

Este notebook detalla la limpieza de los datos previa a la subida en base de datos.

# 2. Importacion y exploracion

## 2.1 Detalles artistas

detalles_artistas = pd.read_csv("../data/raw/detalles_artistas.csv")
detalles_artistas.sample(5)

Exploración del dataset

se.exploracion_dataframe(detalles_artistas)

Se observa que:
- Hay valores erróneos en qué hace, diferentes a las profesiones. Son dificiles de limpiar, así que se dejan así de momento.


- Al menos el 50% de los Actores tienen una fecha de nacimiento igual a 9999, que equivaldría a nulo.
- es_conocido_por tiene nulos marcados en texto; 'No hay datos sobre porqué es conocido'
- de forma similar, premios tiene los 0 marcados como 'no tiene premios'

Corrigiendo estos fallos:

repl_dict = {"conocido_por":{'No hay datos sobre porqué es conocido':np.nan},
            "premios":{'no tiene premios':0},
            "año_nac":{9999:np.nan}
             }
detalles_artistas.replace({"conocido_por":{'No hay datos sobre porqué es conocido':np.nan}}, inplace=True)
detalles_artistas.replace({"premios":{'no tiene premios':0}}, inplace=True)
detalles_artistas.replace({"año_nac":{9999:np.nan}}, inplace=True)
detalles_artistas

Tal vez sea necesario adicionalmente explotar las peliculas por las que cada artista es conocido.

## 2.2 Detalles peliculas

detalles_peliculas = pd.read_csv("../data/raw/detalles_peliculas.csv")
detalles_peliculas.sample(5)

se.exploracion_dataframe(detalles_peliculas)

Parece que aqui la unicas correcciones necesarias serían:
- La conversion de duracion a minutos. Se deja para más tarde, la subida a BDD es más prioritaria.
- El cambio de los valores 'desconocido' a nulo.

detalles_peliculas.replace({"Desconocido":np.nan}, inplace=True)

## 2.3 Pelicula artista

pelicula_artista = pd.read_csv("../data/raw/pelicula_artista.csv")
pelicula_artista.sample(5)

se.exploracion_dataframe(pelicula_artista)

Aquí la nueva codificacion de valor nulo es 'Sin datos'.

pelicula_artista.replace({"Sin datos":np.nan}, inplace=True)

## 2.4 Peliculas

peliculas = pd.read_csv("../data/raw/peliculas.csv")
peliculas.sample(5)

se.exploracion_dataframe(peliculas)

La variable mes tiene 13 valores únicos, así que hay que comprobarlos todos.

peliculas["Mes"].value_counts()

Hay mes 0, así que se asume que es un valor nulo.

peliculas.loc[peliculas["Mes"]==0, "Mes"] = np.nan

También hay años 9999. Se asume que son nulos.

peliculas.loc[peliculas["Año"]==0, "Año"] = np.nan

Se observa que los títulos de las peliculas, que deberían ser únicos, en algunos casos más frecuentes aparecen 2 veces. Conviene realizar otra comprobacion:

peliculas_doble = (peliculas["Titulo"].value_counts()[peliculas["Titulo"].value_counts() > 1]).index.to_list()
peliculas[peliculas["Titulo"].isin(peliculas_doble)].sort_values("Titulo")

Debe tratarse de Remakes o cosas similares, ya que difieren en id y en fecha, así que se dejan tal cual.

# 3. Subida datos limpios a `datos/cleaned`

detalles_artistas.to_csv("../data/cleaned/detalles_artistas.csv")
detalles_peliculas.to_csv("../data/cleaned/detalles_peliculas.csv")
pelicula_artista.to_csv("../data/cleaned/pelicula_artista.csv")
peliculas.to_csv("../data/cleaned/peliculas.csv")
